take out wrp.Message
