+++ eventmetrics/parsers/provide.go
@@ -1,7 +1,6 @@
 package parsers
 
 import (
-	"github.com/go-kit/kit/log"
 	"github.com/xmidt-org/glaukos/eventmetrics/queue"
 	"github.com/xmidt-org/glaukos/message/history"
 	"go.uber.org/fx"
@@ -24,27 +23,6 @@ func provideParsers() fx.Option {
 			Target: func(measures Measures) queue.Parser {
 				return MetadataParser{
 					Measures: measures,
-				}
-			},
-		},
-		fx.Annotated{
-			Group: "parsers",
-			Target: func(logger log.Logger, measures Measures, client *history.CodexClient) queue.Parser {
-				return &BootTimeParser{
-					Measures: measures,
-					Logger:   logger,
-					Client:   client,
-				}
-			},
-		},
-		fx.Annotated{
-			Group: "parsers",
-			Target: func(logger log.Logger, measures Measures, client *history.CodexClient) queue.Parser {
-				return &RebootTimeParser{
-					Measures: measures,
-					Logger:   logger,
-					Client:   client,
-					Label:    "reboot_to_manageable_parser",
 				}
 			},
 		},
