+++ eventmetrics/transport.go
@@ -10,10 +10,12 @@ import (
 	"fmt"
 	"io/ioutil"
 	"net/http"
+	"time"
 
 	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
 	kithttp "github.com/go-kit/kit/transport/http"
+	"github.com/xmidt-org/glaukos/message"
 	"github.com/xmidt-org/themis/xlog"
 	"github.com/xmidt-org/wrp-go/v3"
 )
@@ -66,5 +68,7 @@ func DecodeEvent(_ context.Context, r *http.Request) (interface{}, error) {
 		return nil, fmt.Errorf("could not decode request body: %v", err)
 	}
 
-	return msg, nil
+	event := message.NewEvent(msg, time.Now)
+
+	return event, nil
 }
