+++ message/validation/rule_test.go
@@ -9,9 +9,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
-	"github.com/xmidt-org/glaukos/eventmetrics/queue"
 	"github.com/xmidt-org/glaukos/message"
-	"github.com/xmidt-org/wrp-go/v3"
 )
 
 func TestNewEventValidation(t *testing.T) {
@@ -103,7 +101,7 @@ func TestNewEventValidation(t *testing.T) {
 	}
 }
 
-func TestIsEventValid(t *testing.T) {
+func TestValidEvent(t *testing.T) {
 	now, err := time.Parse(time.RFC3339Nano, "2021-03-02T18:00:01Z")
 	assert.Nil(t, err)
 
@@ -121,12 +119,12 @@ func TestIsEventValid(t *testing.T) {
 		{
 			description: "Valid Event",
 			event: message.Event{
-				MsgType: 4,
-				Dest:    "event:device-status/mac:112233445566/some-event/1613033276/2s",
+				MsgType:     4,
+				Destination: "event:device-status/mac:112233445566/some-event/1613033276/2s",
 				Metadata: map[string]string{
 					message.BootTimeKey: fmt.Sprint(now.Unix()),
 				},
-				BirthDate: now.UnixNano(),
+				Birthdate: now.UnixNano(),
 			},
 			timeLocation: Birthdate,
 			timeIsValid:  true,
@@ -135,8 +133,8 @@ func TestIsEventValid(t *testing.T) {
 		{
 			description: "Wrong Event Type",
 			event: message.Event{
-				MsgType: 4,
-				Dest:    "event:device-status/mac:112233445566/online",
+				MsgType:     4,
+				Destination: "event:device-status/mac:112233445566/online",
 				Metadata: map[string]string{
 					message.BootTimeKey: fmt.Sprint(now.Unix()),
 				},
@@ -147,7 +145,7 @@ func TestIsEventValid(t *testing.T) {
 		{
 			description: "No boot-time",
 			event: message.Event{
-				Dest: "event:device-status/mac:112233445566/some-event/1613033276/2s",
+				Destination: "event:device-status/mac:112233445566/some-event/1613033276/2s",
 			},
 			timeLocation: Boottime,
 			expectedRes:  false,
@@ -156,8 +154,8 @@ func TestIsEventValid(t *testing.T) {
 		{
 			description: "Boot-time Invalid",
 			event: message.Event{
-				Dest:     "event:device-status/mac:112233445566/some-event/1613033276/2s",
-				Metadata: map[string]string{message.BootTimeKey: "not-a-number"},
+				Destination: "event:device-status/mac:112233445566/some-event/1613033276/2s",
+				Metadata:    map[string]string{message.BootTimeKey: "not-a-number"},
 			},
 			timeLocation: Boottime,
 			expectedRes:  false,
@@ -166,8 +164,8 @@ func TestIsEventValid(t *testing.T) {
 		{
 			description: "Boot-time Too Old",
 			event: message.Event{
-				Dest:     "event:device-status/mac:112233445566/some-event/1613033276/2s",
-				Metadata: map[string]string{message.BootTimeKey: "60"},
+				Destination: "event:device-status/mac:112233445566/some-event/1613033276/2s",
+				Metadata:    map[string]string{message.BootTimeKey: "60"},
 			},
 			timeLocation: Boottime,
 			timeIsValid:  false,
@@ -178,7 +176,7 @@ func TestIsEventValid(t *testing.T) {
 		{
 			description: "Birthdate Invalid",
 			event: message.Event{
-				Dest: "event:device-status/mac:112233445566/some-event/1613033276/2s",
+				Destination: "event:device-status/mac:112233445566/some-event/1613033276/2s",
 				Metadata: map[string]string{
 					message.BootTimeKey: fmt.Sprint(now.Unix()),
 				},
@@ -202,130 +200,6 @@ func TestIsEventValid(t *testing.T) {
 				timeValidation: mockTimeVal,
 			}
 			res, err := validation.ValidateEvent(tc.event)
-			if tc.expectedErr == nil || err == nil {
-				assert.Equal(tc.expectedErr, err)
-			} else {
-				assert.True(errors.Is(err, tc.expectedErr),
-					fmt.Errorf("error [%v] doesn't contain error [%v] in its err chain",
-						err, tc.expectedErr),
-				)
-			}
-
-			assert.Equal(tc.expectedRes, res)
-		})
-	}
-}
-
-func TestIsWRPValid(t *testing.T) {
-	now, err := time.Parse(time.RFC3339Nano, "2021-03-02T18:00:01Z")
-	assert.Nil(t, err)
-
-	testRegex := regexp.MustCompile("device-status/.*/some-event/.*")
-
-	tests := []struct {
-		description  string
-		wrp          queue.WrpWithTime
-		timeIsValid  bool
-		timeValError error
-		timeLocation TimeLocation
-		expectedRes  bool
-		expectedErr  error
-	}{
-		{
-			description: "Valid WRP",
-			wrp: queue.WrpWithTime{
-				Message: wrp.Message{
-					Destination: "event:device-status/mac:112233445566/some-event/1613033276/2s",
-					Metadata: map[string]string{
-						message.BootTimeKey: fmt.Sprint(now.Unix()),
-					},
-				},
-				Beginning: now,
-			},
-			timeLocation: Birthdate,
-			timeIsValid:  true,
-			expectedRes:  true,
-		},
-		{
-			description: "Wrong Event Type",
-			wrp: queue.WrpWithTime{
-				Message: wrp.Message{
-					Destination: "event:device-status/mac:112233445566/online",
-					Metadata: map[string]string{
-						message.BootTimeKey: fmt.Sprint(now.Unix()),
-					},
-				},
-			},
-			expectedRes: false,
-			expectedErr: ErrInvalidEventType,
-		},
-		{
-			description: "No Boot-time",
-			wrp: queue.WrpWithTime{
-				Message: wrp.Message{
-					Destination: "event:device-status/mac:112233445566/some-event/1613033276/2s",
-					Metadata:    map[string]string{},
-				},
-			},
-			timeLocation: Boottime,
-			expectedRes:  false,
-			expectedErr:  ErrInvalidBootTime,
-		},
-		{
-			description: "Boot-time Invalid",
-			wrp: queue.WrpWithTime{
-				Message: wrp.Message{
-					Destination: "event:device-status/mac:112233445566/some-event/1613033276/2s",
-					Metadata:    map[string]string{message.BootTimeKey: "not-a-number"},
-				},
-			},
-			timeLocation: Boottime,
-			expectedRes:  false,
-			expectedErr:  ErrInvalidBootTime,
-		},
-		{
-			description: "Boot-time Too Old",
-			wrp: queue.WrpWithTime{
-				Message: wrp.Message{
-					Destination: "event:device-status/mac:112233445566/some-event/1613033276/2s",
-					Metadata:    map[string]string{message.BootTimeKey: "60"},
-				},
-			},
-			timeLocation: Boottime,
-			timeIsValid:  false,
-			timeValError: ErrPastDate,
-			expectedRes:  false,
-			expectedErr:  ErrInvalidBootTime,
-		},
-		{
-			description: "Birthdate Invalid",
-			wrp: queue.WrpWithTime{
-				Message: wrp.Message{
-					Destination: "event:device-status/mac:112233445566/some-event/1613033276/2s",
-					Metadata: map[string]string{
-						message.BootTimeKey: fmt.Sprint(now.Unix()),
-					},
-				},
-			},
-			timeLocation: Birthdate,
-			timeIsValid:  false,
-			timeValError: ErrPastDate,
-			expectedRes:  false,
-			expectedErr:  ErrInvalidBirthdate,
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.description, func(t *testing.T) {
-			assert := assert.New(t)
-			mockTimeVal := new(mockTimeValidation)
-			mockTimeVal.On("IsTimeValid", mock.Anything).Return(tc.timeIsValid, tc.timeValError).Once()
-			validation := rule{
-				regex:          testRegex,
-				calculateUsing: tc.timeLocation,
-				timeValidation: mockTimeVal,
-			}
-			res, err := validation.ValidateWRP(tc.wrp)
 			if tc.expectedErr == nil || err == nil {
 				assert.Equal(tc.expectedErr, err)
 			} else {
@@ -384,7 +258,7 @@ func TestGetEventCompareTime(t *testing.T) {
 			description: "Successful with boot-time",
 			event: message.Event{
 				Metadata:  map[string]string{message.BootTimeKey: "50"},
-				BirthDate: 60,
+				Birthdate: 60,
 			},
 			timeLocation: Boottime,
 			expectedTime: time.Unix(50, 0),
@@ -393,7 +267,7 @@ func TestGetEventCompareTime(t *testing.T) {
 			description: "Successful with birthdate",
 			event: message.Event{
 				Metadata:  map[string]string{message.BootTimeKey: "50"},
-				BirthDate: 60,
+				Birthdate: 60,
 			},
 			timeLocation: Birthdate,
 			expectedTime: time.Unix(0, 60),
@@ -402,7 +276,7 @@ func TestGetEventCompareTime(t *testing.T) {
 			description: "Boot-time doesn't exist",
 			event: message.Event{
 				Metadata:  map[string]string{},
-				BirthDate: 60,
+				Birthdate: 60,
 			},
 			timeLocation: Boottime,
 			expectedTime: time.Time{},
@@ -413,71 +287,7 @@ func TestGetEventCompareTime(t *testing.T) {
 		t.Run(tc.description, func(t *testing.T) {
 			assert := assert.New(t)
 			validation := rule{calculateUsing: tc.timeLocation}
-			time, err := validation.EventTime(tc.event)
-			assert.True(tc.expectedTime.Equal(time))
-			assert.Equal(tc.expectedErr, err)
-		})
-	}
-}
-
-func TestGetWRPCompareTime(t *testing.T) {
-	tests := []struct {
-		description  string
-		wrp          queue.WrpWithTime
-		timeLocation TimeLocation
-		expectedTime time.Time
-		expectedErr  error
-	}{
-		{
-			description: "Successful with boot-time",
-			wrp: queue.WrpWithTime{
-				Message: wrp.Message{
-					Metadata: map[string]string{message.BootTimeKey: "50"},
-				},
-				Beginning: time.Unix(0, 60),
-			},
-			timeLocation: Boottime,
-			expectedTime: time.Unix(50, 0),
-		},
-		{
-			description: "Successful with birthdate",
-			wrp: queue.WrpWithTime{
-				Message: wrp.Message{
-					Metadata: map[string]string{message.BootTimeKey: "50"},
-				},
-				Beginning: time.Unix(0, 60),
-			},
-			timeLocation: Birthdate,
-			expectedTime: time.Unix(0, 60),
-		},
-		{
-			description: "Boot-time doesn't exist",
-			wrp: queue.WrpWithTime{
-				Message: wrp.Message{
-					Metadata: map[string]string{},
-				},
-				Beginning: time.Unix(0, 60),
-			},
-			timeLocation: Boottime,
-			expectedTime: time.Time{},
-			expectedErr:  message.ErrBootTimeNotFound,
-		},
-		{
-			description: "Default Birthdate",
-			wrp: queue.WrpWithTime{
-				Message: wrp.Message{
-					Metadata: map[string]string{},
-				},
-			},
-			timeLocation: Birthdate,
-			expectedTime: time.Time{},
-		},
-	}
-	for _, tc := range tests {
-		t.Run(tc.description, func(t *testing.T) {
-			assert := assert.New(t)
-			validation := rule{calculateUsing: tc.timeLocation}
-			time, err := validation.WRPTime(tc.wrp)
+			time, err := validation.GetCompareTime(tc.event)
 			assert.True(tc.expectedTime.Equal(time))
 			assert.Equal(tc.expectedErr, err)
 		})
